% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addChIP,TrenaMultiScore-method}
\alias{addChIP,TrenaMultiScore-method}
\title{add 1 if a motif hit intersects a ChIP-seq hit, otherwise 0.}
\usage{
\S4method{addChIP}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}

\item{mtx}{a numerical matrix, genes are rownames, samples are colnames}
}
\value{
None
}
\description{
add 1 if a motif hit intersects a ChIP-seq hit, otherwise 0.  full-with ChIP peaks are used.
}
