% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addDistanceToTSS,TrenaMultiScore-method}
\alias{addDistanceToTSS,TrenaMultiScore-method}
\title{append 'TSS' column to the accumulating table}
\usage{
\S4method{addDistanceToTSS}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
None
}
\description{
append 'TSS' column to the accumulating table, recording distance from the main
  transcript's TSS to the start of the motif hit, negative values for upstream locations
}
