% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addGeneExpressionCorrelations,TrenaMultiScore-method}
\alias{addGeneExpressionCorrelations,TrenaMultiScore-method}
\title{add a tf/targetGene correlations score for those TFs also in the expression matrix}
\usage{
\S4method{addGeneExpressionCorrelations}{TrenaMultiScore}(obj, mtx)
}
\arguments{
\item{obj}{a TrenaMultiScore object}

\item{mtx}{a numerical matrix, genes are rownames, samples are colnames}
}
\value{
None
}
\description{
add a tf/targetGene correlations score for those TFs also in the expression matrix
}
