% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addGenicAnnotations,TrenaMultiScore-method}
\alias{addGenicAnnotations,TrenaMultiScore-method}
\title{add intron, exon, utr, promoter, etc annotations}
\usage{
\S4method{addGenicAnnotations}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}

\item{mtx}{a numerical matrix, genes are rownames, samples are colnames}
}
\value{
None
}
\description{
add intron, exon, utr, promoter, etc annotations
}
