% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addRnaBindingProteins,TrenaMultiScore-method}
\alias{addRnaBindingProteins,TrenaMultiScore-method}
\title{query remote khaleesi (ingested POSTAR2) database for rna binding protein binding sites for the target gene}
\usage{
\S4method{addRnaBindingProteins}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
data.frame
}
\description{
POSTAR2 curated many clip-seq (and related) data sources; we put them in a postgres database, queried here by gene
}
