% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{findOpenChromatin,TrenaMultiScore-method}
\alias{findOpenChromatin,TrenaMultiScore-method}
\title{find open chromatin in the current gene region, or (if supplied) somewhere else}
\usage{
\S4method{findOpenChromatin}{TrenaMultiScore}(
  obj,
  chrom = NA,
  start = NA,
  end = NA,
  use.merged.atac = FALSE
)
}
\arguments{
\item{obj}{a TrenaMultiScore object}

\item{chrom}{end, integer, chromosomal location, default NA}
}
\description{
Open chromatin from DNase footprints or scATAC-seq is retrieved, saved for later use
}
