% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{findFimoTFBS,TrenaMultiScore-method}
\alias{findFimoTFBS,TrenaMultiScore-method}
\title{get scored FIMO motif binding sites}
\usage{
\S4method{findFimoTFBS}{TrenaMultiScore}(
  obj,
  motifs = NA,
  fimo.threshold = NA,
  chrom = NA,
  start = NA,
  end = NA
)
}
\arguments{
\item{motifs}{a list of MotifDb items, JASPAR2019 hsapiens by default}

\item{fimo.threshold}{1e-4 by default}

\item{chrom}{geneHancerRegion chrom by default}

\item{start}{geneHancerRegion start by default}

\item{end}{geneHancerRegion end by default}
}
\value{
a data.frame
}
\description{
call FimoBatch to get data.frame of all matches above threshold
}
