% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getGeneHancerRegion,TrenaMultiScore-method}
\alias{getGeneHancerRegion,TrenaMultiScore-method}
\title{return the full extent of GeneHancer regions for the targetGene}
\usage{
\S4method{getGeneHancerRegion}{TrenaMultiScore}(obj)
}
\value{
a data.frame
}
\description{
GeneHancer has distal enhancers withing ~1MB of the gene; return that region
}
