% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getMultiScoreTable,TrenaMultiScore-method}
\alias{getMultiScoreTable,TrenaMultiScore-method}
\title{returns the current state of scored motif matches}
\usage{
\S4method{getMultiScoreTable}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
a data.frame
}
\description{
can be called at any time
}
