% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getOpenChromatin,TrenaMultiScore-method}
\alias{getOpenChromatin,TrenaMultiScore-method}
\title{retrieve currently assigned open chromatin}
\usage{
\S4method{getOpenChromatin}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
data.frame
}
\description{
Return a data.frame with the open chromatin previously retrieved
}
