% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getProject,TrenaMultiScore-method}
\alias{getProject,TrenaMultiScore-method}
\title{return the TMS project}
\usage{
\S4method{getProject}{TrenaMultiScore}(obj)
}
\value{
a TrenaProject instance
}
\description{
get a reference to the TrenaProject concrete subclass included in this object
}
