% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getRnaBindingProteins,TrenaMultiScore-method}
\alias{getRnaBindingProteins,TrenaMultiScore-method}
\title{return previously queried data.frame of rna binding protein binding sites for the target gene}
\usage{
\S4method{getRnaBindingProteins}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
data.frame
}
\description{
POSTAR2 curated many clip-seq (and related) data sources; current value is returned here
}
