% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{getTargetGeneInfo,TrenaMultiScore-method}
\alias{getTargetGeneInfo,TrenaMultiScore-method}
\title{get basic stats on the primary transcript}
\usage{
\S4method{getTargetGeneInfo}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
None
}
\description{
get basic stats on the primary transcript
}
