% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{scoreMotifHitsForOpenChromatin,TrenaMultiScore-method}
\alias{scoreMotifHitsForOpenChromatin,TrenaMultiScore-method}
\alias{scoreMotifHitsForConservation,TrenaMultiScore-method}
\title{add open chromatin intersection info (none, partial, complete for the currently held fimo table}
\usage{
\S4method{scoreMotifHitsForOpenChromatin}{TrenaMultiScore}(obj)

\S4method{scoreMotifHitsForConservation}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
a data.frame

a data.frame
}
\description{
adds "none", "partial", "complete" annotation to the "oc" column for each motif hit in the
  already built (or supplied) fimo table, using the current obj@state$openChromatin table

adds several conservation scores, each an average, to each motif hit in the already build fimo table
}
