% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{scoreMotifHitsForGeneHancer,TrenaMultiScore-method}
\alias{scoreMotifHitsForGeneHancer,TrenaMultiScore-method}
\title{attach a genehancer score to each motif hit, 0 if none}
\usage{
\S4method{scoreMotifHitsForGeneHancer}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
a data.frame
}
\description{
adds several conservation scores, each an average, to each motif hit in the already build fimo table
}
