% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{addGeneExpressionCorrelations,TrenaMultiScore-method}
\alias{addGeneExpressionCorrelations,TrenaMultiScore-method}
\title{add a tf/targetGene correlations score for those TFs also in the expression matrix}
\usage{
\S4method{addGeneExpressionCorrelations}{TrenaMultiScore}(
  obj,
  mtx,
  featureName = "cor",
  colnames = list(),
  method = "pearson"
)
}
\arguments{
\item{obj}{a TrenaMultiScore object}

\item{mtx}{a numerical matrix, genes are rownames, samples are colnames}

\item{featureName}{a character string, default "cor", data.frame column where results are stored}

\item{colnames}{list of character strings, limits scope of correlation calcuation. default empty list: use all columns}

\item{method}{a character string, either "spearman" or "pearson"}
}
\value{
None
}
\description{
add a tf/targetGene correlations score for those TFs also in the expression matrix
}
