% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\name{scoreMotifHitsForConservation,TrenaMultiScore-method}
\alias{scoreMotifHitsForConservation,TrenaMultiScore-method}
\title{add conservation scores to the currently held fimo table}
\usage{
\S4method{scoreMotifHitsForConservation}{TrenaMultiScore}(obj)
}
\arguments{
\item{obj}{a TrenaMultiScore object}
}
\value{
a data.frame
}
\description{
adds several conservation scores, each an average, to each motif hit in the already build fimo table
}
