% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaMultiScore.R
\docType{class}
\name{TrenaMultiScore-class}
\alias{TrenaMultiScore-class}
\alias{.TrenaMultiScore}
\alias{TrenaMultiScore}
\title{TrenaMultiScore-class}
\usage{
TrenaMultiScore(trenaProject, targetGene, quiet = TRUE)
}
\arguments{
\item{trenaProject}{a concrete subclass of TrenaProject}
}
\value{
An object of the TrenaMultiScore class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
